Clean up telemetry code a bit more in ChatService (#259449)
* Simplify telemetry in ChatService

* Use ChatRequestTelemetry
