+++ src/vs/workbench/contrib/chat/common/chatServiceTelemetry.ts
@@ -3,8 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { URI } from '../../../../base/common/uri.js';
+import { isLocation } from '../../../../editor/common/languages.js';
 import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry.js';
-import { ChatAgentVoteDirection, ChatCopyKind, IChatUserActionEvent } from './chatService.js';
+import { IChatAgentData } from './chatAgents.js';
+import { ChatRequestModel, IChatRequestVariableData } from './chatModel.js';
+import { ChatRequestAgentPart, ChatRequestAgentSubcommandPart, ChatRequestSlashCommandPart } from './chatParserTypes.js';
+import { ChatAgentVoteDirection, ChatCopyKind, IChatSendRequestOptions, IChatUserActionEvent } from './chatService.js';
+import { isImageVariableEntry } from './chatVariableEntries.js';
+import { ChatAgentLocation } from './constants.js';
+import { ILanguageModelsService } from './languageModels.js';
 
 type ChatVoteEvent = {
 	direction: 'up' | 'down';
@@ -124,6 +132,44 @@ type ChatEditHunkClassification = {
 	comment: 'Provides insight into the usage of Chat features.';
 };
 
+export type ChatProviderInvokedEvent = {
+	timeToFirstProgress: number | undefined;
+	totalTime: number | undefined;
+	result: 'success' | 'error' | 'errorWithOutput' | 'cancelled' | 'filtered';
+	requestType: 'string' | 'followup' | 'slashCommand';
+	chatSessionId: string;
+	agent: string;
+	agentExtensionId: string | undefined;
+	slashCommand: string | undefined;
+	location: ChatAgentLocation;
+	citations: number;
+	numCodeBlocks: number;
+	isParticipantDetected: boolean;
+	enableCommandDetection: boolean;
+	attachmentKinds: string[];
+	model: string | undefined;
+};
+
+export type ChatProviderInvokedClassification = {
+	timeToFirstProgress: { classification: 'SystemMetaData'; purpose: 'PerformanceAndHealth'; comment: 'The time in milliseconds from invoking the provider to getting the first data.' };
+	totalTime: { classification: 'SystemMetaData'; purpose: 'PerformanceAndHealth'; comment: 'The total time it took to run the provider\'s `provideResponseWithProgress`.' };
+	result: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'Whether invoking the ChatProvider resulted in an error.' };
+	requestType: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The type of request that the user made.' };
+	chatSessionId: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'A random ID for the session.' };
+	agent: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The type of agent used.' };
+	agentExtensionId: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The extension that contributed the agent.' };
+	slashCommand?: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The type of slashCommand used.' };
+	location: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The location at which chat request was made.' };
+	citations: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The number of public code citations that were returned with the response.' };
+	numCodeBlocks: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The number of code blocks in the response.' };
+	isParticipantDetected: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'Whether the participant was automatically detected.' };
+	enableCommandDetection: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'Whether participation detection was disabled for this invocation.' };
+	attachmentKinds: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The types of variables/attachments that the user included with their query.' };
+	model: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The model used to generate the response.' };
+	owner: 'roblourens';
+	comment: 'Provides insight into the performance of Chat agents.';
+};
+
 export class ChatServiceTelemetry {
 	constructor(
 		@ITelemetryService private readonly telemetryService: ITelemetryService,
@@ -184,5 +230,120 @@ export class ChatServiceTelemetry {
 			command,
 			numFollowups,
 		});
+	}
+}
+
+function getCodeBlocks(text: string): string[] {
+	const lines = text.split('\n');
+	const codeBlockLanguages: string[] = [];
+
+	let codeBlockState: undefined | { readonly delimiter: string; readonly languageId: string };
+	for (let i = 0; i < lines.length; i++) {
+		const line = lines[i];
+
+		if (codeBlockState) {
+			if (new RegExp(`^\\s*${codeBlockState.delimiter}\\s*$`).test(line)) {
+				codeBlockLanguages.push(codeBlockState.languageId);
+				codeBlockState = undefined;
+			}
+		} else {
+			const match = line.match(/^(\s*)(`{3,}|~{3,})(\w*)/);
+			if (match) {
+				codeBlockState = { delimiter: match[2], languageId: match[3] };
+			}
+		}
+	}
+	return codeBlockLanguages;
+}
+
+export class ChatRequestTelemetry {
+	private isComplete = false;
+
+	constructor(private readonly opts: {
+		agentPart: ChatRequestAgentPart | undefined;
+		agentSlashCommandPart: ChatRequestAgentSubcommandPart | undefined;
+		commandPart: ChatRequestSlashCommandPart | undefined;
+		sessionId: string;
+		location: ChatAgentLocation;
+		options: IChatSendRequestOptions | undefined;
+		enableCommandDetection: boolean;
+	},
+		@ITelemetryService private readonly telemetryService: ITelemetryService,
+		@ILanguageModelsService private readonly languageModelsService: ILanguageModelsService
+	) { }
+
+	complete({ timeToFirstProgress, totalTime, result, requestType, request, detectedAgent }: {
+		timeToFirstProgress: number | undefined;
+		totalTime: number | undefined;
+		result: ChatProviderInvokedEvent['result'];
+		requestType: ChatProviderInvokedEvent['requestType'];
+		// Should rearrange so these 2 can be in the constructor
+		request: ChatRequestModel;
+		detectedAgent: IChatAgentData | undefined;
+	}) {
+		if (this.isComplete) {
+			return;
+		}
+
+		this.isComplete = true;
+		this.telemetryService.publicLog2<ChatProviderInvokedEvent, ChatProviderInvokedClassification>('interactiveSessionProviderInvoked', {
+			timeToFirstProgress,
+			totalTime,
+			result,
+			requestType,
+			agent: detectedAgent?.id ?? this.opts.agentPart?.agent.id ?? '',
+			agentExtensionId: detectedAgent?.extensionId.value ?? this.opts.agentPart?.agent.extensionId.value ?? '',
+			slashCommand: this.opts.agentSlashCommandPart ? this.opts.agentSlashCommandPart.command.name : this.opts.commandPart?.slashCommand.command,
+			chatSessionId: this.opts.sessionId,
+			enableCommandDetection: this.opts.enableCommandDetection,
+			isParticipantDetected: !!detectedAgent,
+			location: this.opts.location,
+			citations: request.response?.codeCitations.length ?? 0,
+			numCodeBlocks: getCodeBlocks(request.response?.response.toString() ?? '').length,
+			attachmentKinds: this.attachmentKindsForTelemetry(request.variableData),
+			model: this.resolveModelId(this.opts.options?.userSelectedModelId),
+		});
+	}
+
+	private attachmentKindsForTelemetry(variableData: IChatRequestVariableData): string[] {
+		// this shows why attachments still have to be cleaned up somewhat
+		return variableData.variables.map(v => {
+			if (v.kind === 'implicit') {
+				return 'implicit';
+			} else if (v.range) {
+				// 'range' is range within the prompt text
+				if (v.kind === 'tool') {
+					return 'toolInPrompt';
+				} else if (v.kind === 'toolset') {
+					return 'toolsetInPrompt';
+				} else {
+					return 'fileInPrompt';
+				}
+			} else if (v.kind === 'command') {
+				return 'command';
+			} else if (v.kind === 'symbol') {
+				return 'symbol';
+			} else if (isImageVariableEntry(v)) {
+				return 'image';
+			} else if (v.kind === 'directory') {
+				return 'directory';
+			} else if (v.kind === 'tool') {
+				return 'tool';
+			} else if (v.kind === 'toolset') {
+				return 'toolset';
+			} else {
+				if (URI.isUri(v.value)) {
+					return 'file';
+				} else if (isLocation(v.value)) {
+					return 'location';
+				} else {
+					return 'otherAttachment';
+				}
+			}
+		});
+	}
+
+	private resolveModelId(userSelectedModelId: string | undefined): string | undefined {
+		return userSelectedModelId && this.languageModelsService.lookupLanguageModel(userSelectedModelId)?.id;
 	}
 }
